[LBSE] Move the SVGLayerTransformUpdater class in its own file https://bugs.webkit.org/show_bug.cgi?id=242682
Reviewed by Adrian Perez de Castro.

RenderSVGRoot needs to use SVGLayerTransformUpdater as well in future,
so move the class in its own file.

* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/rendering/svg/RenderSVGContainer.cpp:
(WebCore::SVGLayerTransformUpdater::SVGLayerTransformUpdater): Deleted.
(WebCore::SVGLayerTransformUpdater::~SVGLayerTransformUpdater): Deleted.
* Source/WebCore/rendering/svg/SVGLayerTransformUpdater.h: Added.
(WebCore::SVGLayerTransformUpdater::SVGLayerTransformUpdater):
(WebCore::SVGLayerTransformUpdater::~SVGLayerTransformUpdater):

Canonical link: https://commits.webkit.org/252449@main
